[refactor] string_handler diary app으로 이동
