+++ diary/string_handler.py
